Update IniFile.cs
Hey Abhijeet, I figured out the issues and added the corrected code. I liked the way you explained me, and i was not able to sleep after the interview as the code was still running in my mind so i thought of fixing this code first. Thank you for the opportunity.